Refactor: 메인페이지 비즈니스 로직 custom hook으로 분리 #70
