+++ src/components/Home/PopularExperience/PopularExperience.tsx
@@ -1,49 +1,14 @@
-import { useEffect, useRef, useState } from 'react';
 import clsx from 'clsx';
-import { useQuery } from '@tanstack/react-query';
 import { Splide, SplideSlide } from '@splidejs/react-splide';
 import '@splidejs/splide/dist/css/themes/splide-default.min.css';
-import toast from 'react-hot-toast';
 
 import Card from '@/components/Home/Card/Card';
-import { getActivities } from '@/api/activities';
-import QUERY_KEYS from '@/constants/queryKeys';
-import useInfiniteScroll from '@/hooks/common/useInfiniteScroll';
 import LeftArrow from '#/icons/icon-left-arrow.svg';
 import styles from './PopularExperience.module.css';
+import { usePopularExperience } from '@/hooks/Home/usePopularExperience';
 
 function PopularExperience({ deviceType }: { deviceType: string | undefined }) {
-  const [slideIndex, setSlideIndex] = useState(0);
-  const splideRef = useRef(null);
-  const [cursorId, setCursorId] = useState<number | null | undefined>();
-  const { data, refetch, isError } = useQuery({
-    queryKey: [QUERY_KEYS.popularActivities],
-    queryFn: () => getActivities({ method: 'cursor', sort: 'most_reviewed', size: 4, cursorId }),
-  });
-  const [cardData, setCardData] = useState(data?.activities ?? []);
-  const { isVisible, targetRef } = useInfiniteScroll();
-
-  /** 버튼을 통한 slide 함수 */
-  const handleSlideByBtn = (num: number) => {
-    const newIndex = slideIndex + num;
-    if (newIndex < 0 || newIndex + 2 >= cardData.length) return;
-
-    setSlideIndex((prev) => prev + num);
-    if (splideRef.current) {
-      (splideRef.current as any).go(newIndex);
-    }
-  };
-
-  useEffect(() => {
-    if (cursorId === null) return;
-    if (isVisible) refetch();
-  }, [isVisible]);
-
-  useEffect(() => {
-    setCardData((prev) => [...prev, ...(data?.activities ?? [])]);
-    setCursorId(data?.cursorId);
-    if (isError) toast.error('데이터를 불러오지 못했습니다.');
-  }, [data]);
+  const { slideIndex, handleSlideByBtn, setSlideIndex, splideRef, targetRef, cardData } = usePopularExperience();
 
   return (
     <section className={styles.container}>
